fix circular dependency and stories
