+++ client/src/components/composite/SignUpForm/SignUpForm.tsx
@@ -1,12 +1,16 @@
 import PaginatedForm, {
   PageProps
 } from "components/generic/PaginatedForm/PaginatedForm"
-import { PAGES, STEPPER_PROPS } from "./PageConfig/PageConfig"
+import { STEPPER_PROPS } from "./PageConfig/PageConfig"
 import Stepper from "components/generic/StepperComponent/StepperComponent"
 import { oneLevelUp, useCurrentStep } from "./utils/Utils"
 import { useAppData } from "store/store"
 import { Navigate } from "react-router-dom"
-import { PAYMENT_INFORMATION_ROUTE, PERSONAL_ROUTE_1 } from "./utils/RouteNames"
+import {
+  PAGES,
+  PAYMENT_INFORMATION_ROUTE,
+  PERSONAL_ROUTE_1
+} from "./utils/RouteNames"
 
 interface ISignUpFormProps {
   currentPage: PAGES
